RAG-Final/vertex_ai_retriever.py
from google.cloud import aiplatform
from typing import List, Dict, Any

class VertexAIRetriever:
    def __init__(self, project_id: str, location: str, index_id: str):
        self.project_id = project_id
        self.location = location
        self.index_id = index_id
        aiplatform.init(project=project_id, location=location)
        
        # Get the index
        self.index = aiplatform.MatchingEngineIndex(index_name=self.index_id)
        
        # Get or create the index endpoint
        index_endpoints = aiplatform.MatchingEngineIndexEndpoint.list(
            filter=f'display_name="{self.index_id}-endpoint"'
        )
        if index_endpoints:
            self.index_endpoint = index_endpoints[0]
        else:
            self.index_endpoint = aiplatform.MatchingEngineIndexEndpoint.create(
                display_name=f"{self.index_id}-endpoint",
                project=project_id,
                location=location
            )
        
        # Deploy the index to the endpoint if not already deployed
        deployments = self.index_endpoint.list_deployed_indexes()
        if not any(d.index == self.index.resource_name for d in deployments):
            self.index_endpoint.deploy_index(index=self.index)

    def upsert_document(self, document: Dict[str, Any]):
        # Implement upsert logic here
        # You'll need to use the appropriate Vertex AI SDK calls to upsert the document
        print(f"Upserting document with ID: {document['id']}")
        # Example (you may need to adjust this based on your exact requirements):
        # self.index.upsert_datapoints(
        #     embeddings=[document['content']],  # Assuming 'content' is the text to be embedded
        #     ids=[document['id']],
        #     feature_vector_lengths=[len(document['content'])],
        # )

    def clear_index(self):
        # Implement clear index logic here
        # You'll need to use the appropriate Vertex AI SDK calls to clear the index
        print(f"Clearing index: {self.index_id}")
        # Example (you may need to adjust this based on your exact requirements):
        # self.index.remove_datapoints(datapoint_ids=None)  # Remove all datapoints

    # Add any other necessary methods for your use case
